simple_dnns-main/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=2.116=openblas
blas-devel=3.9.0=16_osx64_openblas
brotli=1.0.9=h5eb16cf_7
brotli-bin=1.0.9=h5eb16cf_7
brotlipy=0.7.0=py310h1961e1f_1004
bzip2=1.0.8=h0d85af4_4
ca-certificates=2022.6.15=h033912b_0
certifi=2022.6.15=pyhd8ed1ab_1
cffi=1.15.1=py310h96bbf6e_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
cryptography=37.0.4=py310h52c3658_0
cycler=0.11.0=pyhd8ed1ab_0
fonttools=4.36.0=py310h3c08dca_0
freetype=2.12.1=h3f81eb7_0
idna=3.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=hac89ed1_2
kiwisolver=1.4.4=py310habb735a_0
lcms2=2.12=h577c468_0
lerc=4.0.0=hb486fe8_0
libblas=3.9.0=16_osx64_openblas
libbrotlicommon=1.0.9=h5eb16cf_7
libbrotlidec=1.0.9=h5eb16cf_7
libbrotlienc=1.0.9=h5eb16cf_7
libcblas=3.9.0=16_osx64_openblas
libcxx=14.0.6=hce7ea42_0
libdeflate=1.13=h775f41a_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=10_4_0_h97931a8_25
libgfortran5=11.3.0=h082f757_25
liblapack=3.9.0=16_osx64_openblas
liblapacke=3.9.0=16_osx64_openblas
libopenblas=0.3.21=openmp_h947e540_2
libpng=1.6.37=h5481273_4
libprotobuf=3.20.1=hfa58983_1
libsqlite=3.39.2=h5a3d3bf_1
libtiff=4.4.0=h5e0c7b4_3
libwebp-base=1.2.4=h775f41a_0
libxcb=1.13=h0d85af4_1004
libzlib=1.2.12=hfe4f2af_2
llvm-openmp=14.0.4=ha654fa7_0
matplotlib=3.5.3=py310h2ec42d9_1
matplotlib-base=3.5.3=py310h6438751_1
mkl=2022.1.0=h860c996_928
mkl-devel=2022.1.0=h694c41f_929
mkl-include=2022.1.0=h6bab518_928
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h96cf925_1
ninja=1.11.0=h1b54a9f_0
numpy=1.23.2=py310hf910466_0
openblas=0.3.21=openmp_h2bb6d6c_2
openjpeg=2.5.0=h5d0d7b0_1
openssl=1.1.1q=hfe4f2af_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.4=py310hecf8f37_0
patsy=0.5.2=pyhd8ed1ab_0
pillow=9.2.0=py310h54af1cc_2
pip=22.2.2=pyhd8ed1ab_0
pthread-stubs=0.4=hc929b4f_1001
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py310h2ec42d9_5
python=3.10.6=ha7b0be1_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.12.1=cpu_py310h32957a8_0
pytz=2022.2.1=pyhd8ed1ab_0
readline=8.1.2=h3899abd_0
requests=2.28.1=pyhd8ed1ab_0
scikit-learn=1.1.2=py310h3ba92a3_0
scipy=1.9.1=py310h240c617_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
setuptools=65.2.0=py310h2ec42d9_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h6db0672_2
statsmodels=0.13.2=py310h1bbcd0e_0
tbb=2021.5.0=hbb4e6a2_1
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h5dbffcc_0
torchvision=0.13.0=cpu_py310h87c54a7_0
tornado=6.2=py310h6c45266_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022c=h191b570_0
unicodedata2=14.0.0=py310h1961e1f_1
urllib3=1.26.11=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h35c211d_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
zstd=1.5.2=hfa58983_4
